Initial commit for main EC2 instance
This starts a small ARM64 EC2 instance based on the latest Ubuntu 22.10,
and deploys it in the Default VPC region on a random public subnet.
It's initialized with a cloud-init configuration that'll be generated at
build-time. Ports are opened for both HTTP and HTTPS access, as well as
SSH access. Note: the SSH Key needs to be generated and downloaded
manually from the AWS console and stored on the local machine.